code clean up
git-svn-id: https://svn.apache.org/repos/asf/chemistry/opencmis/trunk@1695691 13f79535-47bb-0310-9956-ffa450edef68

+++ chemistry-opencmis-commons/chemistry-opencmis-commons-impl/src/main/java/org/apache/chemistry/opencmis/commons/impl/MimeHelper.java
@@ -523,7 +523,7 @@ public final class MimeHelper {
                     // self-defining token type.
                 } else if (c < 32 || c >= 127 || delimiters.indexOf(c) != -1) {
                     pos++;
-                    return new Token((int) c, String.valueOf(c));
+                    return new Token(c, String.valueOf(c));
                 } else {
                     // start of an atom, parse it off.
                     return readAtomicToken();
